Build: added the --environment option to the build script.
This option can be used to force the build script to use the native gcc
or the the gcc used by the ANaConDA framework. The gcc which is used by
the ANaConDA framework should always be able to compile the LWM library.
Default is to use the native gcc compiler, which might fail to compile
the LWM library.